Improve PHP + State DB in progress
*Update php code, pour enlever les erreurs dans le log
*mysql connection set to UTF8
*UI: Ajout du state manager
*Php pour saver state manager on DB

+++ php/classes/Mails.php
@@ -15,7 +15,7 @@ class Mails {
 		
 		$query = "SELECT * FROM mails " ;
 		//Conditions
-		if ($params->id)
+		if (isset($params->id) && $params->id)
 		{
 			$query .= ' WHERE id = '.$params->id;
 		}
@@ -37,7 +37,7 @@ class Mails {
 	function delete(stdClass $params)
 	{
 		//Conditions
-		if ($params->id)
+		if (isset($params->id) && $params->id)
 		{
 			$query = "DELETE FROM mails " ;
 			$query .= ' WHERE sent = 0 AND id = '.$params->id;
@@ -88,7 +88,7 @@ class Mails {
 	
 	    return $results;	
 	}
-	function getMailFromSousTraitant(stdClass $params)
+	function getMailFromSousTraitant($params)
 	{
 		$sousTraitantId = $params->filter[0]->value;
